# frontend/lib/api/openapi.json
{
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "Page_ReportSummary_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/ReportSummary"
      },
      "title": "Items",
      "type": "array"
     },
     "page": {
      "title": "Page",
      "type": "integer"
     },
     "page_size": {
      "title": "Page Size",
      "type": "integer"
     },
     "total": {
      "title": "Total",
      "type": "integer"
     }
    },
    "required": [
     "items",
     "total",
     "page",
     "page_size"
    ],
    "title": "Page[ReportSummary]",
    "type": "object"
   },
...
   "ReportDetail": {
    "properties": {
     "compare_run_ids": {
      "items": {
       "type": "string"
      },
      "title": "Compare Run Ids",
      "type": "array"
     },
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "include_narrative": {
      "title": "Include Narrative",
      "type": "boolean"
     },
     "run_id": {
      "title": "Run Id",
      "type": "string"
     },
     "sections": {
      "items": {
       "type": "string"
      },
      "title": "Sections",
      "type": "array"
     },
     "snapshot": {
      "additionalProperties": true,
      "description": "Frozen copy computed at creation; never recomputed",
      "title": "Snapshot",
      "type": "object"
     },
     "snapshot_hash": {
      "title": "Snapshot Hash",
      "type": "string"
     },
     "spec": {
      "$ref": "#/components/schemas/ReportSpec"
     },
     "title": {
      "title": "Title",
      "type": "string"
     }
    },
    "required": [
     "id",
     "title",
     "run_id",
     "sections",
     "compare_run_ids",
     "include_narrative",
     "snapshot_hash",
     "created_at",
     "spec",
     "snapshot"
    ],
    "title": "ReportDetail",
    "type": "object"
   },
   "ReportSpec": {
    "properties": {
     "compare_run_ids": {
      "description": "Runs compared with the main run",
      "items": {
       "type": "string"
      },
      "maxItems": 3,
      "title": "Compare Run Ids",
      "type": "array"
     },
     "include_narrative": {
      "default": false,
      "title": "Include Narrative",
      "type": "boolean"
     },
     "run_id": {
      "description": "Main run of the report",
      "maxLength": 64,
      "title": "Run Id",
      "type": "string"
     },
     "sections": {
      "items": {
       "enum": [
        "summary",
        "financial",
        "operational",
        "buyers",
        "logistics",
        "crops",
        "insights",
        "comparison"
       ],
       "type": "string"
      },
      "minItems": 1,
      "title": "Sections",
      "type": "array"
     },
     "title": {
      "maxLength": 160,
      "minLength": 1,
      "title": "Title",
      "type": "string"
     }
    },
    "required": [
     "title",
     "run_id"
    ],
    "title": "ReportSpec",
    "type": "object"
   },
   "ReportSummary": {
    "properties": {
     "compare_run_ids": {
      "items": {
       "type": "string"
      },
      "title": "Compare Run Ids",
      "type": "array"
     },
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "include_narrative": {
      "title": "Include Narrative",
      "type": "boolean"
     },
     "run_id": {
      "title": "Run Id",
      "type": "string"
     },
     "sections": {
      "items": {
       "type": "string"
      },
      "title": "Sections",
      "type": "array"
     },
     "snapshot_hash": {
      "title": "Snapshot Hash",
      "type": "string"
     },
     "title": {
      "title": "Title",
      "type": "string"
     }
    },
    "required": [
     "id",
     "title",
     "run_id",
     "sections",
     "compare_run_ids",
     "include_narrative",
     "snapshot_hash",
     "created_at"
    ],
    "title": "ReportSummary",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "Harvest2Value API",
  "version": "2.0.0-dev"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v2/reports": {
   "get": {
    "operationId": "list_reports_api_v2_reports_get",
    "parameters": [
     {
      "in": "query",
      "name": "run_id",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Run Id"
      }
     },
     {
      "in": "query",
      "name": "page",
      "required": false,
      "schema": {
       "default": 1,
       "minimum": 1,
       "title": "Page",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "page_size",
      "required": false,
      "schema": {
       "default": 20,
       "maximum": 100,
       "minimum": 1,
       "title": "Page Size",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_ReportSummary_"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Reports",
    "tags": [
     "reports"
    ]
   },
   "post": {
    "description": "Computes every chosen section once and freezes it: later edits of data, scenarios or runs\nnever change the report.",
    "operationId": "create_report_api_v2_reports_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ReportSpec"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReportDetail"
        }
       }
      },
      "description": "Successful Response"
     },
     "404": {
      "description": "Run not found"
     },
     "409": {
      "description": "The main run has no plan"
     },
     "422": {
      "description": "Invalid spec"
     }
    },
    "summary": "Create Report",
    "tags": [
     "reports"
    ]
   }
  },
...
  "/api/v2/reports/{report_id}": {
   "delete": {
    "operationId": "delete_report_api_v2_reports__report_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "report_id",
      "required": true,
      "schema": {
       "title": "Report Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "404": {
      "description": "Report not found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete Report",
    "tags": [
     "reports"
    ]
   },
   "get": {
    "operationId": "get_report_api_v2_reports__report_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "report_id",
      "required": true,
      "schema": {
       "title": "Report Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReportDetail"
        }
       }
      },
      "description": "Successful Response"
     },
     "404": {
      "description": "Report not found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Report",
    "tags": [
     "reports"
    ]
   }
  },
  "/api/v2/reports/{report_id}/export.csv": {
   "get": {
    "operationId": "export_csv_api_v2_reports__report_id__export_csv_get",
    "parameters": [
     {
      "in": "path",
      "name": "report_id",
      "required": true,
      "schema": {
       "title": "Report Id",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "section",
      "required": true,
      "schema": {
       "title": "Section",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "table",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Table"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       },
       "text/csv": {}
      },
      "description": "Successful Response"
     },
     "404": {
      "description": "Report, section or table not found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Export Csv",
    "tags": [
     "reports"
    ]
   }
  },
  "/api/v2/reports/{report_id}/export.json": {
   "get": {
    "operationId": "export_json_api_v2_reports__report_id__export_json_get",
    "parameters": [
     {
      "in": "path",
      "name": "report_id",
      "required": true,
      "schema": {
       "title": "Report Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "404": {
      "description": "Report not found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Export Json",
    "tags": [
     "reports"
    ]
   }
  },
...
 }
}